// generated/src/app/api-models/procedure-unit-value.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProcedureUnitValue {
  procedureCode: string;
  seqUnitValue: number;
  scaleId: string;
  geographicRegion: string;
  effectiveDate: Date;
  totalUnits: number;
  totalUnitsConvFt: string;
  professionalUnits: number;
  profUnitsConvFt: string;
  technicalUnits: number;
  techUnitsConvFt: string;
  anesthesiaUnits: number;
  anesUnitsConvFt: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}